project table 컴포넌트
